+++ 42cursus/libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: haekang <haekang@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/22 14:41:43 by haekang           #+#    #+#             */
-/*   Updated: 2023/03/27 03:30:26 by haekang          ###   ########.fr       */
+/*   Updated: 2023/04/09 17:20:36 by haekang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,5 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 		i++;
 	}
 	dst[i] = '\0';
-	return ((size_t)ft_strlen((char *)src));
+	return (ft_strlen((char *)src));
 }
